+++ src/app.cpp
@@ -1,4 +1,4 @@
-#include "app.hpp"
+#include "../include/app.hpp"
 
 //
 // App
