+++ src/cloud.cpp
@@ -1,4 +1,4 @@
-#include "cloud.hpp"
+#include "../include/cloud.hpp"
 
 //
 // PointCyl
